HDDS-12300. Merge OmUpgradeConfig into OmConfig
